+++ gen_libs.py
@@ -581,7 +581,7 @@ def dict_2_list(dict_list, key_val, **kwargs):
 
     """
 
-    return [row[key_val] for row in list(dict_list)]
+    return [row[key_val] for row in list(dict_list) if key_val in row]
 
 
 def dict_2_std(data, ofile=False, **kwargs):